feat(ui): rename settings
