--- src/ui/components/Setting/index.ts
@@ -1,1 +1,0 @@
-export * from "./Setting";
