+++ src/ui/components/Settings/components/ChangePin/ChangePin.scss
