+++ src/ui/components/Settings/components/RecoverySeedPhrase/ConfirmModal.tsx
