+++ src/ui/components/Settings/components/RecoverySeedPhrase/RecoverySeedPhrase.test.tsx
