+++ src/ui/components/Settings/components/RecoverySeedPhrase/RecoverySeedPhrase.tsx
