+++ src/ui/components/Settings/components/RecoverySeedPhrase/RecoverySeedPhrase.types.ts
