+++ src/ui/components/Settings/components/RecoverySeedPhrase/index.ts
