+++ src/ui/components/Settings/components/SettingList.scss
