+++ src/ui/components/Settings/components/SettingsItem.tsx
@@ -1,7 +1,7 @@
 import { IonIcon, IonItem, IonLabel, IonNote } from "@ionic/react";
 import { chevronForward } from "ionicons/icons";
 import "./SettingsItem.scss";
-import { SettingsItemProps } from "../Setting.types";
+import { SettingsItemProps } from "../Settings.types";
 
 const SettingsItem = ({ item, handleOptionClick }: SettingsItemProps) => {
   const content = (
