+++ src/ui/components/Settings/components/TermsAndPrivacy/TermsAndPrivacy.test.tsx
