+++ src/ui/components/Settings/components/TermsAndPrivacy/TermsAndPrivacy.tsx
