+++ src/ui/components/Settings/components/TermsAndPrivacy/index.ts
